requirements.txt
faiss-gpu==1.6.5
h5py==2.10.0
matplotlib==3.5.3
numpy==1.20.3
opencv-python-headless==4.12.0.88
Pillow==9.4.0
scikit-learn==0.22.1
scipy==1.7.3
timm==0.5.4
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.67.1
yacs==0.1.8
